# make application config cacheable & cache it in production