--- a/config/application.config.php
+++ b/config/application.config.php
@@ -69,5 +69,24 @@ return array(
             './module',
             './vendor',
         ),
+
+        // Whether or not to enable a configuration cache.
+        // If enabled, the merged configuration will be cached and used in
+        // subsequent requests.
+        'config_cache_enabled' => ('production' == getenv('APPLICATION_ENV')),
+
+        // The key used to create the configuration cache file name.
+        'config_cache_key' => 'zf2.configuration',
+
+        // Whether or not to enable a module class map cache.
+        // If enabled, creates a module class map cache which will be used
+        // by in future requests, to reduce the autoloading process.
+        'module_map_cache_enabled' => ('production' == getenv('APPLICATION_ENV')),
+
+        // The key used to create the class map cache file name.
+        'module_map_cache_key' => 'zf2.classmap',
+
+        // The path in which to cache merged configuration.
+        'cache_dir' => 'data/cache',
     ),
 );