--- a/module/CommonBundle/Component/Authentication/Factory/Adapter/Doctrine.php
+++ b/module/CommonBundle/Component/Authentication/Factory/Adapter/Doctrine.php
@@ -16,13 +16,24 @@
  * @license http://litus.cc/LICENSE
  */
 
-return array(
-    'service_manager' => array(
-        'factories' => array(
-            'cache' =>
-                ('production' == getenv('APPLICATION_ENV'))
-                ? 'CommonBundle\Component\ApplicationConfig\Cache\Memcached'
-                : 'CommonBundle\Component\ApplicationConfig\Cache\Memory',
-        ),
-    ),
-);
+namespace CommonBundle\Component\Authentication\Factory\Adapter;
+
+use CommonBundle\Component\Authentication\Adapter\Doctrine\Credential as DoctrineAdapter,
+    Zend\ServiceManager\ServiceLocatorInterface;
+
+/**
+ * Factory to create a Doctrine Adapter.
+ *
+ * @author Bram Gotink <bram.gotink@litus.cc>
+ */
+class Doctrine implements \Zend\ServiceManager\FactoryInterface
+{
+    public function createService(ServiceLocatorInterface $sl)
+    {
+        return new DoctrineAdapter(
+            $sl->get('doctrine.entitymanager.orm_default'),
+            'CommonBundle\Entity\User\Person',
+            'username'
+        );
+    }
+}