--- a/module/CommonBundle/Component/Form/Factory/Bootstrap.php
+++ b/module/CommonBundle/Component/Form/Factory/Bootstrap.php
@@ -16,13 +16,17 @@
  * @license http://litus.cc/LICENSE
  */
 
-return array(
-    'service_manager' => array(
-        'factories' => array(
-            'cache' =>
-                ('production' == getenv('APPLICATION_ENV'))
-                ? 'CommonBundle\Component\ApplicationConfig\Cache\Memcached'
-                : 'CommonBundle\Component\ApplicationConfig\Cache\Memory',
-        ),
-    ),
-);
+namespace CommonBundle\Component\Form\Factory;
+
+/**
+ * A factory to create a FormFactory for bootstrap forms.
+ *
+ * @author Bram Gotink <bram.gotink@litus.cc>
+ */
+class Bootstrap extends AbstractFactory
+{
+    public function __construct()
+    {
+        parent::__construct('bootstrap');
+    }
+}