--- a/module/SyllabusBundle/Component/XMLParser/Study.php
+++ b/module/SyllabusBundle/Component/XMLParser/Study.php
@@ -476,7 +476,7 @@ class Study
 
                 $generalGroups = $this->getGeneralMandatoryGroups($combination['entity']->getPhase(), $moduleGroups);
                 if (!empty($generalGroups)) {
-                   $groups = array_merge($groups, $generalGroups);
+                    $groups = array_merge($groups, $generalGroups);
                 }
             }
 
@@ -485,19 +485,20 @@ class Study
     }
 
     /**
-     * @param  int $phase
+     * @param  int   $phase
      * @param  array $moduleGroups
      * @return array
      */
-    private function getGeneralMandatoryGroups($phase, $moduleGroups) {
+    private function getGeneralMandatoryGroups($phase, $moduleGroups)
+    {
         $groups = array();
 
         //for each child node in the same phase check if branch is fully mandatory
-        foreach($moduleGroups as $group) {
-            if($phase == $group->getPhase() && count($group->getChildren()) == 0) {
+        foreach ($moduleGroups as $group) {
+            if ($phase == $group->getPhase() && count($group->getChildren()) == 0) {
                 if ($this->isFullMandatoryBranch($group)) {
                     $groups[] = $group;
-                } else {
+                }
             }
         }
 
@@ -508,18 +509,19 @@ class Study
      * @param  moduleGroup $group
      * @return boolean
      */
-    private function isFullMandatoryBranch($group) {
+    private function isFullMandatoryBranch($group)
+    {
         $result = false;
-        if($group->isMandatory()){
+        if ($group->isMandatory()) {
             if (null !== $group->getParent()) {
                 $result = $this->isFullMandatoryBranch($group->getParent());
             } else {
                 return true;
             }
         }
+
         return $result;
     }
-
 
     /**
      * @return EntityManager