Cache the application config.

**What's this about?**

ZF2 loads the configuration for every request. That means interpreting over 125 PHP files, executing them and merging the resulting arrays into one. ZF2 can cache the resulting mega-array, skipping the entire load/merge process.

**What's the catch?**

To allow for caching of the configuration, the configuration _mustn't_ contain anything other than arrays and primitives (strings, booleans, numbers). Functions are not allowed. This means all functions in our configuration had to be rewritten into factory classes.

This is not suited for development, and the cached configuration should explicitly be cleared in any deploy process.

**Is it useful?**

To test this, I removed the `->run()` part of `public/index.html` and ran `ab -n 200 -a 10 http://localhost:54887/`.

without caching:

```
This is ApacheBench, Version 2.3 <$Revision: 1638069 $>
Copyright 1996 Adam Twiss, Zeus Technology Ltd, http://www.zeustech.net/
Licensed to The Apache Software Foundation, http://www.apache.org/

Benchmarking 127.0.0.1 (be patient)
Completed 100 requests
Completed 200 requests
Finished 200 requests


Server Software:        nginx/1.4.6
Server Hostname:        127.0.0.1
Server Port:            54887

Document Path:          /
Document Length:        16 bytes

Concurrency Level:      10
Time taken for tests:   22.498 seconds
Complete requests:      200
Failed requests:        19
   (Connect: 0, Receive: 0, Length: 19, Exceptions: 0)
Total transferred:      29178 bytes
HTML transferred:       3178 bytes
Requests per second:    8.89 [#/sec] (mean)
Time per request:       1124.905 [ms] (mean)
Time per request:       112.491 [ms] (mean, across all concurrent requests)
Transfer rate:          1.27 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        0    0   0.1      0       1
Processing:   145 1109 145.4   1102    1419
Waiting:      145 1108 145.4   1102    1419
Total:        146 1109 145.3   1102    1419

Percentage of the requests served within a certain time (ms)
  50%   1102
  66%   1128
  75%   1152
  80%   1193
  90%   1298
  95%   1333
  98%   1387
  99%   1418
 100%   1419 (longest request)
```

with caching:

```
This is ApacheBench, Version 2.3 <$Revision: 1638069 $>
Copyright 1996 Adam Twiss, Zeus Technology Ltd, http://www.zeustech.net/
Licensed to The Apache Software Foundation, http://www.apache.org/

Benchmarking 127.0.0.1 (be patient)
Completed 100 requests
Completed 200 requests
Finished 200 requests


Server Software:        nginx/1.4.6
Server Hostname:        127.0.0.1
Server Port:            54887

Document Path:          /
Document Length:        16 bytes

Concurrency Level:      10
Time taken for tests:   18.249 seconds
Complete requests:      200
Failed requests:        19
   (Connect: 0, Receive: 0, Length: 19, Exceptions: 0)
Total transferred:      29179 bytes
HTML transferred:       3179 bytes
Requests per second:    10.96 [#/sec] (mean)
Time per request:       912.456 [ms] (mean)
Time per request:       91.246 [ms] (mean, across all concurrent requests)
Transfer rate:          1.56 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        0    0   0.1      0       1
Processing:   194  899  88.9    909    1139
Waiting:      194  899  88.9    909    1139
Total:        195  899  88.8    909    1139

Percentage of the requests served within a certain time (ms)
  50%    909
  66%    919
  75%    926
  80%    932
  90%    940
  95%    950
  98%    980
  99%   1083
 100%   1139 (longest request)
```