@bgotink 2015-09-13
Seems like someone forgot to run `vendor/bin/php-cs-fixer fix .`